+++ src/renderer/src/store/slices/settingsSlice.ts
@@ -15,6 +15,7 @@ export const defaultGeneralSettings: GeneralSettings = {
   warnWhenSwitchingThemes: true,
   warnWhenEditingSnippet: true,
   warnWhenCloningSnippet: true,
+  warnWhenClickingReadonlySnippet: true,
   codeEditorTheme: 'default',
   codeEditorSetup: { ...DEFAULT_CODE_EDITOR_SETUP },
   codeEditorFontSize: DEFAULT_CODE_EDITOR_FONT_SIZE,