+++ src/renderer/src/ui/Settings/constants.ts
@@ -53,6 +53,7 @@ export const DEFAULT_GENERAL_SETTINGS: GeneralSettings = {
   warnWhenSwitchingThemes: true,
   warnWhenEditingSnippet: true,
   warnWhenCloningSnippet: true,
+  warnWhenClickingReadonlySnippet: true,
   codeEditorTheme: 'default',
   codeEditorSetup: { ...DEFAULT_CODE_EDITOR_SETUP },
   codeEditorFontSize: DEFAULT_CODE_EDITOR_FONT_SIZE,